fix: Fix Analyze button that was waiting end of worker process even though model was already available.
* Rename Analyze Audio to Bake Audio for clarity

+++ Core/LIPSYNC2D_VoskHelper.py
@@ -280,7 +280,7 @@ class LIPSYNC2D_VoskHelper():
         return
 
     @staticmethod
-    def check_worker_finished() -> int | None:
+    def check_worker_finished() -> float | None:
         """
         Checks the state of the worker process for LIPSYNC2D_VoskHelper and updates
         related addon preferences if necessary.
